pyproject.toml
[project]
name = "textagents"
version = "0.0.1"
description = "Declarative, file-based PydanticAI agents"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
        {name = "Jan Siml", email = "49557684+svilupp@users.noreply.github.com"},
]
keywords = ["llm", "pydantic", "agents", "evaluation", "judge", "pydantic-ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-ai>=0.1.0",
    "python-dotenv>=1.2.1",
    "textprompts>=0.1.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
logfire = ["logfire>=0.1.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ty",
    "ruff>=0.1.0",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.27",
]

[project.scripts]
textagents = "textagents.cli:app"

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ty.environment]
python-version = "3.11"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.coverage.run]
source = ["src/textagents"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
